+++ packages/web/src/components/remix-card/RemixCard.tsx
@@ -1,8 +1,8 @@
 import { ID, Remix } from '@audius/common/models'
 
 import { ArtistPopover } from 'components/artist/ArtistPopover'
-import TrackFlair from 'components/co-sign/TrackFlair'
-import { Size } from 'components/co-sign/types'
+import TrackFlair from 'components/track-flair/TrackFlair'
+import { Size } from 'components/track-flair/types'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
 import UserBadges from 'components/user-badges/UserBadges'
