+++ packages/web/src/components/track/mobile/TrackTileArt.tsx
@@ -3,8 +3,8 @@ import { memo } from 'react'
 import { SquareSizes, ID, Remix } from '@audius/common/models'
 import cn from 'classnames'
 
-import TrackFlair from 'components/co-sign/TrackFlair'
-import { Size } from 'components/co-sign/types'
+import TrackFlair from 'components/track-flair/TrackFlair'
+import { Size } from 'components/track-flair/types'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
 import { useCollectionCoverArt } from 'hooks/useCollectionCoverArt'
 import { useTrackCoverArt } from 'hooks/useTrackCoverArt'
